Move ProjectDetails to its own component
